requirements.txt
faster-whisper>=1.1.0
ffmpeg-python>=0.2.0
tqdm>=4.66.0

